+++ lib/common/htmltable.c
@@ -32,6 +32,7 @@
 #include <common/render.h>
 #include <common/htmltable.h>
 #include <cgraph/agxbuf.h>
+#include <cgraph/prisize_t.h>
 #include <common/pointset.h>
 #include <common/intset.h>
 #include <cdt/cdt.h>
@@ -39,7 +40,9 @@
 #include <cgraph/exit.h>
 #include <cgraph/strcasecmp.h>
 #include <cgraph/unreachable.h>
+#include <float.h>
 #include <inttypes.h>
+#include <math.h>
 #include <stddef.h>
 #include <stdbool.h>
 #include <stdio.h>
@@ -852,7 +855,7 @@ static void free_html_tbl(htmltbl_t * tbl)
 {
     htmlcell_t **cells;
 
-    if (tbl->row_count == -1) {
+    if (tbl->row_count == SIZE_MAX) { // raw, parsed table
 	dtclose(tbl->u.p.rows);
     } else {
 	cells = tbl->u.n.cells;
@@ -1208,8 +1211,8 @@ static int processTbl(graph_t * g, htmltbl_t * tbl, htmlenv_t * env)
     htmlcell_t **cells;
     Dt_t *rows = tbl->u.p.rows;
     int rv = 0;
-    int n_rows = 0;
-    int n_cols = 0;
+    size_t n_rows = 0;
+    size_t n_cols = 0;
     PointSet *ps = newPS();
     Dt_t *is = openIntSet();
 
@@ -1279,32 +1282,32 @@ static void sizeLinearArray(htmltbl_t * tbl)
 {
     htmlcell_t *cp;
     htmlcell_t **cells;
-    int wd, ht, i, x, y;
+    int i;
 
-    tbl->heights = N_NEW(tbl->row_count + 1, int);
-    tbl->widths = N_NEW(tbl->column_count + 1, int);
+    tbl->heights = N_NEW(tbl->row_count + 1, double);
+    tbl->widths = N_NEW(tbl->column_count + 1, double);
 
     for (cells = tbl->u.n.cells; *cells; cells++) {
 	cp = *cells;
+	double ht;
 	if (cp->rowspan == 1)
 	    ht = cp->data.box.UR.y;
 	else {
 	    ht = SPLIT(cp->data.box.UR.y, cp->rowspan, tbl->data.space);
-	    ht = MAX(ht, 1);
+	    ht = fmax(ht, 1);
 	}
+	double wd;
 	if (cp->colspan == 1)
 	    wd = cp->data.box.UR.x;
 	else {
 	    wd = SPLIT(cp->data.box.UR.x, cp->colspan, tbl->data.space);
-	    wd = MAX(wd, 1);
+	    wd = fmax(wd, 1);
 	}
 	for (i = cp->row; i < cp->row + cp->rowspan; i++) {
-	    y = tbl->heights[i];
-	    tbl->heights[i] = MAX(y, ht);
+	    tbl->heights[i] = fmax(tbl->heights[i], ht);
 	}
 	for (i = cp->col; i < cp->col + cp->colspan; i++) {
-	    x = tbl->widths[i];
-	    tbl->widths[i] = MAX(x, wd);
+	    tbl->widths[i] = fmax(tbl->widths[i], wd);
 	}
     }
 }
@@ -1383,12 +1386,11 @@ static void makeGraphs(htmltbl_t * tbl, graph_t * rowg, graph_t * colg)
     node_t *t;
     node_t *lastn;
     node_t *h;
-    int i;
     agxbuf value_buffer = {0};
 
     lastn = NULL;
-    for (i = 0; i <= tbl->column_count; i++) {
-	agxbprint(&value_buffer, "%d", i);
+    for (size_t i = 0; i <= tbl->column_count; i++) {
+	agxbprint(&value_buffer, "%" PRISIZE_T, i);
 	t = agnode(colg, agxbuse(&value_buffer), 1);
 	agbindrec(t, "Agnodeinfo_t", sizeof(Agnodeinfo_t), true);
 	alloc_elist(tbl->row_count, ND_in(t));
@@ -1401,8 +1403,8 @@ static void makeGraphs(htmltbl_t * tbl, graph_t * rowg, graph_t * colg)
 	}
     }
     lastn = NULL;
-    for (i = 0; i <= tbl->row_count; i++) {
-	agxbprint(&value_buffer, "%d", i);
+    for (size_t i = 0; i <= tbl->row_count; i++) {
+	agxbprint(&value_buffer, "%" PRISIZE_T, i);
 	t = agnode(rowg, agxbuse(&value_buffer), 1);
 	agbindrec(t, "Agnodeinfo_t", sizeof(Agnodeinfo_t), true);
 	alloc_elist(tbl->column_count, ND_in(t));
@@ -1486,8 +1488,8 @@ static void sizeArray(htmltbl_t * tbl)
 	return;
     }
 
-    tbl->heights = N_NEW(tbl->row_count + 1, int);
-    tbl->widths = N_NEW(tbl->column_count + 1, int);
+    tbl->heights = N_NEW(tbl->row_count + 1, double);
+    tbl->widths = N_NEW(tbl->column_count + 1, double);
 
     rowg = agopen("rowg", dir, NULL);
     colg = agopen("colg", dir, NULL);
@@ -1677,8 +1679,7 @@ static void pos_html_cell(htmlcell_t * cp, boxf pos, int sides)
  */
 static void pos_html_tbl(htmltbl_t * tbl, boxf pos, int sides)
 {
-    int x, y, delx, dely, oldsz;
-    int i, extra, plus;
+    int plus;
     htmlcell_t **cells = tbl->u.n.cells;
     htmlcell_t *cp;
     boxf cbox;
@@ -1687,12 +1688,10 @@ static void pos_html_tbl(htmltbl_t * tbl, boxf pos, int sides)
 	&& !tbl->data.pencolor)
 	tbl->data.pencolor = gv_strdup(tbl->u.n.parent->data.pencolor);
 
-    oldsz = tbl->data.box.UR.x;
-    delx = pos.UR.x - pos.LL.x - oldsz;
-    assert(delx >= 0);
+    double oldsz = tbl->data.box.UR.x;
+    double delx = fmax(pos.UR.x - pos.LL.x - oldsz, 0);
     oldsz = tbl->data.box.UR.y;
-    dely = pos.UR.y - pos.LL.y - oldsz;
-    assert(dely >= 0);
+    double dely = fmax(pos.UR.y - pos.LL.y - oldsz, 0);
 
     /* If fixed, align box */
     if (tbl->data.flags & FIXED_FLAG) {
@@ -1731,19 +1730,21 @@ static void pos_html_tbl(htmltbl_t * tbl, boxf pos, int sides)
     }
 
     /* change sizes to start positions and distribute extra space */
-    x = pos.LL.x + tbl->data.border + tbl->data.space;
-    extra = delx / tbl->column_count;
-    plus = ROUND(delx - extra * tbl->column_count);
-    for (i = 0; i <= tbl->column_count; i++) {
-	delx = tbl->widths[i] + extra + (i < plus ? 1 : 0);
+    double x = pos.LL.x + tbl->data.border + tbl->data.space;
+    assert(tbl->column_count <= DBL_MAX);
+    double extra = delx / (double)tbl->column_count;
+    plus = ROUND(delx - extra * (double)tbl->column_count);
+    for (size_t i = 0; i <= tbl->column_count; i++) {
+	delx = tbl->widths[i] + extra + ((i <= INT_MAX && (int)i < plus) ? 1 : 0);
 	tbl->widths[i] = x;
 	x += delx + tbl->data.space;
     }
-    y = pos.UR.y - tbl->data.border - tbl->data.space;
-    extra = dely / tbl->row_count;
-    plus = ROUND(dely - extra * tbl->row_count);
-    for (i = 0; i <= tbl->row_count; i++) {
-	dely = tbl->heights[i] + extra + (i < plus ? 1 : 0);
+    double y = pos.UR.y - tbl->data.border - tbl->data.space;
+    assert(tbl->row_count <= DBL_MAX);
+    extra = dely / (double)tbl->row_count;
+    plus = ROUND(dely - extra * (double)tbl->row_count);
+    for (size_t i = 0; i <= tbl->row_count; i++) {
+	dely = tbl->heights[i] + extra + ((i <= INT_MAX && (int)i < plus) ? 1 : 0);
 	tbl->heights[i] = y;
 	y -= dely + tbl->data.space;
     }
@@ -1779,7 +1780,6 @@ static int
 size_html_tbl(graph_t * g, htmltbl_t * tbl, htmlcell_t * parent,
 	      htmlenv_t * env)
 {
-    int i, wd, ht;
     int rv = 0;
     static textfont_t savef;
 
@@ -1798,11 +1798,13 @@ size_html_tbl(graph_t * g, htmltbl_t * tbl, htmlcell_t * parent,
 
     sizeArray(tbl);
 
-    wd = (tbl->column_count + 1) * tbl->data.space + 2 * tbl->data.border;
-    ht = (tbl->row_count + 1) * tbl->data.space + 2 * tbl->data.border;
-    for (i = 0; i < tbl->column_count; i++)
+    assert(tbl->column_count <= DBL_MAX);
+    double wd = ((double)tbl->column_count + 1) * tbl->data.space + 2 * tbl->data.border;
+    assert(tbl->row_count <= DBL_MAX);
+    double ht = ((double)tbl->row_count + 1) * tbl->data.space + 2 * tbl->data.border;
+    for (size_t i = 0; i < tbl->column_count; i++)
 	wd += tbl->widths[i];
-    for (i = 0; i < tbl->row_count; i++)
+    for (size_t i = 0; i < tbl->row_count; i++)
 	ht += tbl->heights[i];
 
     if (tbl->data.flags & FIXED_FLAG) {
@@ -1811,15 +1813,16 @@ size_html_tbl(graph_t * g, htmltbl_t * tbl, htmlcell_t * parent,
 		agerr(AGWARN, "table size too small for content\n");
 		rv = 1;
 	    }
-	    wd = ht = 0;
+	    wd = 0;
+	    ht = 0;
 	} else {
 	    agerr(AGWARN,
 		  "fixed table size with unspecified width or height\n");
 	    rv = 1;
 	}
     }
-    tbl->data.box.UR.x = MAX(wd, tbl->data.width);
-    tbl->data.box.UR.y = MAX(ht, tbl->data.height);
+    tbl->data.box.UR.x = fmax(wd, tbl->data.width);
+    tbl->data.box.UR.y = fmax(ht, tbl->data.height);
 
     if (tbl->font)
 	popFontInfo(env, &savef);
@@ -1932,7 +1935,7 @@ void printTbl(htmltbl_t * tbl, int ind)
 {
     htmlcell_t **cells = tbl->u.n.cells;
     indent(ind);
-    fprintf(stderr, "tbl (%p) %d %d ", tbl, tbl->column_count, tbl->row_count);
+    fprintf(stderr, "tbl (%p) %" PRISIZE_T " %" PRISIZE_T " ", tbl, tbl->column_count, tbl->row_count);
     printData(&tbl->data);
     fputs("\n", stderr);
     while (*cells)