+++ CHANGELOG.md
@@ -41,6 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   [node shape](https://graphviz.org/doc/info/shapes.html#polygon).
 - Second periphery of a cylinder shaped node is not correct. #2297
 - Graphs with more than 127 layers no longer cause out of bound writes. #2355
+- htmltable.c assertions are no longer triggered by using HTML table cells too
+  small to fit their content. #1622
 
 ### Removed